+++ stacks-signer/src/config.rs
@@ -54,6 +54,8 @@ const DEFAULT_PROPOSAL_WAIT_TIME_FOR_PARENT_SECS: u64 = 15;
 /// Default time (in secs) to wait between updating our local state
 /// machine view point and capitulating to other signers tenure view
 const DEFAULT_CAPITULATE_MINER_VIEW_SECS: u64 = 20;
+/// Default HTTP timeout (in seconds) for read/write operations with StackerDB.
+pub const DEFAULT_STACKERDB_TIMEOUT_SECS: u64 = 60;
 
 #[derive(thiserror::Error, Debug)]
 /// An error occurred parsing the provided configuration
@@ -191,6 +193,8 @@ pub struct SignerConfig {
     pub validate_with_replay_tx: bool,
     /// Time to wait between updating our local state machine view point and capitulating to other signers miner view
     pub capitulate_miner_view_timeout: Duration,
+    /// The HTTP timeout for read/write operations with StackerDB.
+    pub stackerdb_timeout: Duration,
     #[cfg(any(test, feature = "testing"))]
     /// Only used for testing purposes to enable overriding the signer version
     pub supported_signer_protocol_version: u64,
@@ -249,6 +253,8 @@ pub struct GlobalConfig {
     pub validate_with_replay_tx: bool,
     /// Time to wait between updating our local state machine view point and capitulating to other signers miner view
     pub capitulate_miner_view_timeout: Duration,
+    /// The HTTP timeout for read/write operations with StackerDB.
+    pub stackerdb_timeout: Duration,
     #[cfg(any(test, feature = "testing"))]
     /// Only used for testing to enable specific signer protocol versions
     pub supported_signer_protocol_version: u64,
@@ -305,6 +311,8 @@ struct RawConfigFile {
     pub validate_with_replay_tx: Option<bool>,
     /// Time to wait (in secs) between updating our local state machine view point and capitulating to other signers miner view
     pub capitulate_miner_view_timeout_secs: Option<u64>,
+    /// Time to wait (in secs) before timing out an HTTP request with StackerDB.
+    pub stackerdb_timeout_secs: Option<u64>,
     #[cfg(any(test, feature = "testing"))]
     /// Only used for testing to enable specific signer protocol versions
     pub supported_signer_protocol_version: Option<u64>,
@@ -439,6 +447,11 @@ impl TryFrom<RawConfigFile> for GlobalConfig {
                 .unwrap_or(DEFAULT_CAPITULATE_MINER_VIEW_SECS),
         );
 
+        let stackerdb_timeout = Duration::from_secs(
+            raw_data
+                .stackerdb_timeout_secs
+                .unwrap_or(DEFAULT_STACKERDB_TIMEOUT_SECS),
+        );
         #[cfg(any(test, feature = "testing"))]
         let supported_signer_protocol_version = raw_data
             .supported_signer_protocol_version
@@ -467,6 +480,7 @@ impl TryFrom<RawConfigFile> for GlobalConfig {
             proposal_wait_for_parent_time,
             validate_with_replay_tx,
             capitulate_miner_view_timeout,
+            stackerdb_timeout,
             #[cfg(any(test, feature = "testing"))]
             supported_signer_protocol_version,
         })